+++ lib/hl_tcl/pkgIndex.tcl
@@ -1,5 +1,5 @@
 
-package ifneeded hl_tcl 0.9.49 [list source [file join $dir hl_tcl.tcl]]
+package ifneeded hl_tcl 1.0.0 [list source [file join $dir hl_tcl.tcl]]
 
 # short intro (for Ruff! docs generator)
 
@@ -117,13 +117,13 @@ The code below:
 
 sets an example of *hl_tcl* usage. Here are the details:
 
-  * **`::stub`** is a procedure to watch the text editing; here it simply puts out the text's last index;
+  * **::stub** is a procedure to watch the text editing; here it simply puts out the text's last index;
 
-  * **`hl_init`** is called *before* filling the text widget with a Tcl code; it sets the highlighting options and disables the highlighting till *hl_text* runs;
+  * **hl_init** is called *before* filling the text widget with a Tcl code; it sets the highlighting options and disables the highlighting till *hl_text* runs;
 
-  * **`hl_text`** runs to highlight the Tcl code of the text widget and to view/edit it.
+  * **hl_text** runs to highlight the Tcl code of the text widget and to view/edit it.
 
-The **`hl_init`** takes arguments:
+The **hl_init** takes arguments:
 
    * *txt* is the text widget's path
    * *args* contains options of text widget (omittable)
@@ -140,14 +140,19 @@ The *args* is a list of *-option "value"* where *-option* may be:
    * *-seen* - number of first lines seen at start (default 500)
    * *-optRE* - flag "use a regular expression to highlight options" (default "yes")
    * *-keywords* - additional commands to highlight (as Tk ones)
+   * *-dobind* - if `true`, forces keys binding at repeating calls of hl_init
+   * *-plaintext* - `true` for plain texts with no highlighting
+   * *-plaincom* - a command for plain highlighting line by line
 
 **Note**: `-seen 500` and `-multiline no` can improve the performance a lot. It's recommended to use `-seen 500` (or any other reasonable limit, e.g. `-seen 200`) at any rate, except for static html pages.
 
+A command for *-plaincom* option has two arguments: a current text's path and a current line's number. It should highlight the current line and return `true`, otherwise (if the current line is Tcl code) it returns `false`. An example of its usage is presented by [alited editor](https://github.com/aplsimple/alited) (`lib/addon` directory).
+
 The rest of *hl_tcl* procedures are:
 
-   *  **`hl_all `** updates all highlighted existing text widgets, e.g. at changing a color scheme of application
-   *  **`hl_readonly`** gets/sets a read-only mode and/or a command to watch a text widget at viewing/editing it
-   *  **`hl_colors`** gets a list of colors for highlighting
+   *  **hl_all** updates all highlighted existing text widgets, e.g. at changing a color scheme of application
+   *  **hl_readonly** gets/sets a read-only mode and/or a command to watch a text widget at viewing/editing it
+   *  **hl_colors** gets a list of colors for highlighting
 
 See details in [Reference](https://aplsimple.github.io/en/tcl/hl_tcl/hl_tcl.html).
 
@@ -192,7 +197,6 @@ The tag pairs can be multiple if the html pages contain them, e.g.
 
 Note that [hl_tcl](https://aplsimple.github.io/en/tcl/hl_tcl/hl_tcl.html) is still disposed to update.
   }
-
 }
 
 namespace eval ::hl_tcl::my {